+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added SlowEffectsOnTimer
+
+### Changed
+
+- Reworked Proximity Indicator
+
 ## [0.5.0] - 2024-05-25
 
 ### Added